feat(AcceptanceTest): base AcceptanceTest 추가
