+++ src/test/java/nextstep/subway/line/LineAcceptanceTest.java
@@ -7,35 +7,19 @@ import static nextstep.subway.line.LineAcceptanceTestFixture.updateLine;
 import static nextstep.subway.station.StationAcceptanceTestFixture.createStation;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
+import nextstep.subway.base.AcceptanceTest;
 import nextstep.subway.helper.JsonPathExtractor;
-import nextstep.subway.helper.TestIsolator;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.web.server.LocalServerPort;
 
 @DisplayName("지하철 노선 관련 기능")
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-public class LineAcceptanceTest {
-    @LocalServerPort
-    int port;
-    @Autowired
-    private TestIsolator testIsolator;
-
-    @BeforeEach
-    public void setUp() {
-        if (RestAssured.port == RestAssured.UNDEFINED_PORT) {
-            RestAssured.port = port;
-        }
-        testIsolator.run();
-    }
+public class LineAcceptanceTest extends AcceptanceTest {
 
     /**
      *  When 지하철 노선을 생성하면
